requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
bsdiff4==1.2.6
certifi==2025.7.14; python_version >= '3.7'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
exceptiongroup==1.3.0; python_version >= '3.7'
fastapi==0.116.1; python_version >= '3.8'
greenlet==3.2.3; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
psutil==5.9.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.4; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.41; python_version >= '3.7'
starlette==0.47.1; python_version >= '3.9'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
watchfiles==1.1.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
